requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.32.0
 